// FRONT/tpi-teg-grupo-13/src/app/core/enums/turn-phase.ts
// Turn phase enums
export enum TurnPhase {
  DEPLOYMENT = 'DEPLOYMENT',
  REINFORCEMENT_5="REINFORCEMENT_5",
  REINFORCEMENT_3="REINFORCEMENT_3",
  REINFORCEMENT = "REINFORCEMENT",  // Colocar ejércitos
  ATTACK = "ATTACK",         // Atacar
  FORTIFY = "FORTIFY",        // Reagrupar
  CLAIM_CARD = "CLAIM_CARD",     // Reclamar carta
  END_TURN = "END_TURN"       // Finalizar turno
}