// FRONT/tpi-teg-grupo-13/src/app/core/enums/event-type.ts
// Game event type enums
export enum EventType {
  GAME_START = 'GAME_START',
  GAME_END = 'GAME_END',
  CONQUEST = 'CONQUEST',
  ATTACK = 'ATTACK',
  REINFORCEMENT = 'REINFORCEMENT',
  FORTIFICATION = 'FORTIFICATION',
  PLAYER_ELIMINATED = 'PLAYER_ELIMINATED',
  NEW_TURN = 'NEW_TURN',
  TRADE_CARDS = 'TRADE_CARDS',
  PLAYER_LEFT = 'PLAYER_LEFT',
  PLAYER_RECONNECTED = 'PLAYER_RECONNECTED',
  OBJECTIVE_COMPLETED = 'OBJECTIVE_COMPLETED'
}